requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic>=2.11.7
supabase==2.18.1
python-dotenv==1.0.0
mangum==0.17.0
